[Scenario] Recreate timeline when top-level activities overlap
When moving activities within a timeline view using mouse drag
controls, the user can cause two activities to overlap. This is
permissable, because top-level activities of a timeline are
allowed occur concurrently; however, this looks strange as the
visual components occupy the same space on screen.

The timeline view previously addressed this possibility only
on view instantiation: Extra 'blocks' were created such that
concurrent activities could be placed on lower rows. With
this change, this behavior may now occur whenever a change to
an activity is made, by recreating the majority of the view
(as would happen during instantiation.) For performance and
useability reasons, this may only occur when a drag gesture
has completed (meaning that activities may still overlap
while a drag is ongoing). Additionally, a check is performed
to detect overlapping activities, to avoid re-creating the
view if it is not necessary.

This addresses nasa/MCT-Plugins#27