[Scenario] Maintain pan on update
Don't update centerTime (which corresponds to
pan) on updateMasterDuration in timeline local
controls. This update is not necessary when
zoomed out (existing clamping of center time
handles expansion of maximum/minimum times)
and causes current pan to be lost if zoomed
in.

Since updateMasterDuration is triggered after
activities are moved, this change avoids causing
activities to jump around visually after drag
gestures complete.