[Scenario] Update timeline view when stale
In nasa/MCT-Plugins#56, user creates a new
Activity in the Directory area and expects
it to show up in the Center Pane (which
is not consistent with what should generally
happen, as views are expected to refresh
only on demand)

To meet this expectation in the special
case of the timeline view, listen for
staleness upon the objects in the
timeline view and explicitly refresh
when this occurs. This is similar to
the approach used in node views.

An extra class (a MergeHandler) is added
to supplement this process. Assuming that
the user does not want to lose unsaved
changes to start/end times, this handler
stores these specific changes and applies
them to incoming versions of objects
before the refresh completes.