[Scenario] Move stale listener to superclass
Move the stale listener used to refresh
Timeline views with incoming changes up
to the superclass shared with Scenario
view, such that both Scenario and
Timeline update with the same strategy.
This allows, for instance, the cost
graph shown in the scenario view to
stay up-to-date with changes to
individual timelines.

This supports resolution of remaining
instances of nasa/MCT-Plugins#56