[Scenario] Add Sub-activity component
Add new component class for Sub-activity in
support of nasa/MCT-Plugins#57. A Sub-activity
is simply an Activity within another Activity,
so it will likely not need any class-specific
behavior; class identity is only used to
distinguish it for purposes of displaying
the correct name to the user and employing
appropriate policy.