[Scenario] Track and save activities changed by constraints
Keep track of any activities whose start and end times
are changed as a result of constraint system
enforcement during an activity dragging gesture.
This ensures that these components are marked dirty
and tracked as such in order to support subsequent
Save All invocations. Failure to do this results in
nasa/MCT-Plugins#92.