[Scenario] Show confirm dialog if export file exists
Prompt the user with a confirmation dialog
when attempting to export a CSV to a file
which already exists, in order to avoid
destroying existing data.

Note that JOptionPane is used instead of
WindowManager here; the latter is generally
preferred to avoid introducing Swing
dependencies around dialogs, but we
already have a Swing dependency here since
the dialog is being launched from a
JFileChooser.