[Scenario] Add dialog for tag selection
Adds a dialog for selecting tag associations for an
activity, as described in nasa/MCT-Plugins#129.

Additionally, refactors the manner in which repository
information is communicated down to custom visual
controls for tag editing. Specifically, repositories
are identified explicitly, as are component type infos
for repository contents; this replaces prior approach,
which identified both repositories and relevant
contents by capability. (This was more dynamic, but
less explicit, and sometimes error prone.)

This change incidentally resolves nasa/MCT-Plugins#132.

This change will additionally simplify
nasa/MCT-Plugins#134, by removing the assumption (in
the code for the visual control in Activity's Info View)
that repositories will be located at the bootstrap
level.